+++ db/migrate/20130522135017_remove_status_from_concepts.rb
@@ -1,0 +1,9 @@
+class RemoveStatusFromConcepts < ActiveRecord::Migration
+  def up
+    remove_column :concepts, :status
+  end
+
+  def down
+    add_column :concepts, :status, :string
+  end
+end